# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "My First Api",
        "description": "",
        "contact": {
            "name": "A2-LAB",
            "url": "http://a2-lab.com/"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/v1/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "description": "Get all users",
                "consumes": [
                    "application/xml",
                    "application/json"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, user data provided"
                    },
                    "400": {
                        "description": "Request errors"
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get user by ID",
                "description": "Get user by ID",
                "consumes": [
                    "application/xml",
                    "application/json"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "type": "number",
                        "default": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, user data provided"
                    },
                    "400": {
                        "description": "Request errors"
                    }
                }
            }
        },
        "/api/v1/users/create": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Add user",
                "description": "Add user",
                "consumes": [
                    "application/xml",
                    "application/json"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "Your name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Your last name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "age",
                        "in": "query",
                        "description": "Your age",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Your email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Your password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, user created"
                    },
                    "400": {
                        "description": "Request errors"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/v1/users/{id}/update": {
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Update user by ID",
                "description": "Update user by ID",
                "consumes": [
                    "application/xml",
                    "application/json"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "type": "number",
                        "default": 1
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "Your name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Your last name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "age",
                        "in": "query",
                        "description": "Your age",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Your email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Your password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, user data provided"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/v1/users/{id}/delete": {
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Delete user by ID",
                "description": "Delete user by ID",
                "consumes": [
                    "application/xml",
                    "application/json"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "type": "number",
                        "default": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, user data provided"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "Error": {
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "tags": [
        {
            "name": "users",
            "description": "CRUD"
        }
    ]
}